project folder structure rework, buttons, world-scale rework

+++ src/main/java/edu/mscd/thesis/util/Util.java
@@ -4,12 +4,12 @@ import edu.mscd.thesis.model.Pos2D;
 
 public class Util {
 
-	public static boolean isValidPos2D(Pos2D p, int xMax, int yMax) {
+	public static boolean isValidPos2D(Pos2D p, double xMax, double yMax) {
 		if (p == null) {
 			return false;
 		} else if (xMax < 1 || yMax < 1) {
 			return false;
-		} else if (p.getX()<0 || p.getY()<0){
+		} else if (p.getX() < 0 || p.getY() < 0) {
 			return false;
 		} else {
 			return p.getX() < xMax && p.getY() < yMax;
